requirements.txt
flask
flask_cors
gunicorn
pillow
colorama
python-magic
Flask-Limiter
flask_session
SQLAlchemy
bcrypt
